chore: upgrade municipio
Customizer: Fixed an issue where dynamic theme settings were not properly loading in the customizer. This ensures that imported settings from the design library are correctly applied in the customizer context.

Favicons: Removes the custom favicon implementation, and migrates it to be done natively in WordPress using the ”site identity” section in the customizer.

Icons: A optimization on the icon rendering has been made. This reduces the size of the material icons bundle up to 85%. This change also brings settings for choosing font style and weight.

General: Minor adjustments to correct an issue where the width settings were not being applied.